% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils.R
\name{qlg_advance_days}
\alias{qlg_advance_days}
\title{Advance a QuantLib date by calendar days}
\usage{
qlg_advance_days(calendar_obj, date_obj, n_days)
}
\arguments{
\item{calendar_obj}{QuantLib calendar object.}

\item{date_obj}{ISO date string, R Date, or QuantLib Date.}

\item{n_days}{Number of days.}
}
\value{
A QuantLib Date object.
}
\description{
Advance a QuantLib date by calendar days
}
