% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_apply_fixings}
\alias{qlg_apply_fixings}
\title{Apply fixing data to cashflow schedule}
\usage{
qlg_apply_fixings(cashflows, fixings, evaluation_date = NULL)
}
\arguments{
\item{cashflows}{A cashflow schedule tibble.}

\item{fixings}{A fixing tibble with fixing_date and fixing columns.}

\item{evaluation_date}{Evaluation date.}
}
\value{
A tibble with fixing_value and is_fixing_known.
}
\description{
Apply fixing data to cashflow schedule
}
