% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_approx_gradient}
\alias{qlg_approx_gradient}
\title{Approximate an objective gradient}
\usage{
qlg_approx_gradient(parameters, objective_function, epsilon = 1e-08)
}
\arguments{
\item{parameters}{Parameter vector.}

\item{objective_function}{Objective function.}

\item{epsilon}{Finite-difference step.}
}
\value{
Numeric gradient.
}
\description{
Approximate an objective gradient
}
