% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_quantive_riemann.R
\name{qlg_asset_swap_analysis}
\alias{qlg_asset_swap_analysis}
\title{Asset swap analysis for a fixed-rate bond}
\usage{
qlg_asset_swap_analysis(
  bond,
  clean_price,
  ibor_index,
  spread,
  settlement_date,
  maturity_date,
  calendar,
  floating_schedule_frequency,
  payment_convention,
  date_generation,
  end_of_month = FALSE,
  floating_day_counter,
  pay_fixed_rate = TRUE,
  par_asset_swap = TRUE,
  discount_curve_handle,
  gearing = 1,
  non_par_repayment = 100
)
}
\description{
Asset swap analysis for a fixed-rate bond
}
