% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory.R
\name{qlg_asset_swap_fair_clean_price}
\alias{qlg_asset_swap_fair_clean_price}
\title{AssetSwap fair clean price}
\usage{
qlg_asset_swap_fair_clean_price(asset_swap)
}
\arguments{
\item{asset_swap}{QuantLib AssetSwap object.}
}
\value{
Numeric fair clean price.
}
\description{
AssetSwap fair clean price
}
