% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory.R
\name{qlg_asset_swap_fair_spread}
\alias{qlg_asset_swap_fair_spread}
\title{AssetSwap fair spread}
\usage{
qlg_asset_swap_fair_spread(asset_swap)
}
\arguments{
\item{asset_swap}{QuantLib AssetSwap object.}
}
\value{
Numeric fair spread.
}
\description{
AssetSwap fair spread
}
