% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory.R
\name{qlg_asset_swap_set_engine}
\alias{qlg_asset_swap_set_engine}
\title{Attach a discounting engine to an AssetSwap}
\usage{
qlg_asset_swap_set_engine(asset_swap, discount_curve_handle)
}
\arguments{
\item{asset_swap}{QuantLib AssetSwap object.}

\item{discount_curve_handle}{QuantLib YieldTermStructureHandle.}
}
\value{
The original AssetSwap object, invisibly.
}
\description{
Attach a discounting engine to an AssetSwap
}
