% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory.R
\name{qlg_asset_swap_summary}
\alias{qlg_asset_swap_summary}
\title{Summarise an AssetSwap}
\usage{
qlg_asset_swap_summary(asset_swap)
}
\arguments{
\item{asset_swap}{QuantLib AssetSwap object.}
}
\value{
A tibble with AssetSwap analytics.
}
\description{
Summarise an AssetSwap
}
