% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_black_calculator_table}
\alias{qlg_black_calculator_table}
\title{Build a Black calculator metric table}
\usage{
qlg_black_calculator_table(
  payoff,
  forward,
  volatility,
  maturity,
  discount_factor
)
}
\arguments{
\item{payoff}{QuantLib payoff.}

\item{forward}{Forward price.}

\item{volatility}{Volatility.}

\item{maturity}{Time to maturity.}

\item{discount_factor}{Discount factor.}
}
\value{
Metric tibble.
}
\description{
Build a Black calculator metric table
}
