% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_black_process_bundle}
\alias{qlg_black_process_bundle}
\title{Build a Black process bundle}
\usage{
qlg_black_process_bundle(
  valuation_date,
  forward,
  risk_free_rate,
  volatility,
  day_counter = QuantLib::Actual365Fixed(),
  calendar = QuantLib::NullCalendar()
)
}
\arguments{
\item{valuation_date}{QuantLib Date.}

\item{forward}{Forward or futures price.}

\item{risk_free_rate}{Risk-free rate.}

\item{volatility}{Black volatility.}

\item{day_counter}{QuantLib day counter.}

\item{calendar}{QuantLib calendar.}
}
\value{
List with process and term structures.
}
\description{
Build a Black process bundle
}
