% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_bond_clean_price_from_yield_safe}
\alias{qlg_bond_clean_price_from_yield_safe}
\title{Safely calculate a bond clean price from yield}
\usage{
qlg_bond_clean_price_from_yield_safe(
  bond,
  yield_rate,
  day_counter,
  compounding,
  frequency,
  settlement_date = NULL
)
}
\arguments{
\item{bond}{QuantLib bond.}

\item{yield_rate}{Yield.}

\item{day_counter}{QuantLib day counter.}

\item{compounding}{Compounding convention.}

\item{frequency}{Coupon frequency.}

\item{settlement_date}{Optional settlement date.}
}
\value{
Numeric clean price or NA.
}
\description{
Safely calculate a bond clean price from yield
}
