% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_bond_convexity}
\alias{qlg_bond_convexity}
\title{Bond Convexity}
\usage{
qlg_bond_convexity(
  bond,
  ytm = NULL,
  day_counter = Actual365Fixed(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\description{
Bond Convexity
}
