% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_bond_coupon_info}
\alias{qlg_bond_coupon_info}
\title{Get bond coupon information}
\usage{
qlg_bond_coupon_info(bond, as_of = qlg_eval_date_get())
}
\arguments{
\item{bond}{A QuantLib bond object.}

\item{as_of}{Evaluation date. Default is  qlg_eval_date_get().}
}
\value{
A tibble with coupon cashflow information.
}
\description{
Get bond coupon information
}
