% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_bond_duration}
\alias{qlg_bond_duration}
\title{Bond Duration}
\usage{
qlg_bond_duration(
  bond,
  ytm = NULL,
  day_counter = Actual365Fixed(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\description{
Bond Duration
}
