% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_bond_futures_ctd_table}
\alias{qlg_bond_futures_ctd_table}
\title{Calculate a Treasury futures CTD table}
\usage{
qlg_bond_futures_ctd_table(
  deliverable_tbl,
  settlement_date,
  futures_price,
  repo_end_date,
  repo_rate,
  settlement_days = 1L,
  calendar = QuantLib::UnitedStates("GovernmentBond"),
  day_counter = QuantLib::ActualActual("Bond"),
  compounding = QuantLib::Compounding_Compounded_get(),
  frequency = QuantLib::Frequency_Semiannual_get(),
  repo_day_counter = QuantLib::Actual360(),
  carry_day_counter = QuantLib::Actual360()
)
}
\arguments{
\item{deliverable_tbl}{Deliverable basket data.}

\item{settlement_date}{Settlement date.}

\item{futures_price}{Futures price.}

\item{repo_end_date}{Repo end date.}

\item{repo_rate}{Repo rate.}

\item{settlement_days}{Settlement days.}

\item{calendar}{QuantLib calendar.}

\item{day_counter}{Bond day counter.}

\item{compounding}{Compounding convention.}

\item{frequency}{Coupon frequency.}

\item{repo_day_counter}{Repo day counter.}

\item{carry_day_counter}{Carry day counter.}
}
\value{
List with basket and CTD row.
}
\description{
Calculate a Treasury futures CTD table
}
