% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_bond_npv_with_curve}
\alias{qlg_bond_npv_with_curve}
\title{Bond NPV with a curve handle}
\usage{
qlg_bond_npv_with_curve(bond, curve_handle)
}
\arguments{
\item{bond}{QuantLib bond object.}

\item{curve_handle}{QuantLib YieldTermStructureHandle.}
}
\value{
Numeric NPV.
}
\description{
Bond NPV with a curve handle
}
