% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_bond_price_from_yield}
\alias{qlg_bond_price_from_yield}
\title{Bond Price from Yield}
\usage{
qlg_bond_price_from_yield(
  bond,
  ytm,
  day_counter = Actual360(),
  compounding = "Compounded",
  frequency = "Annual",
  settlement_date = NULL
)
}
\description{
Bond Price from Yield
}
