% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_bond_price_yield_curve}
\alias{qlg_bond_price_yield_curve}
\title{Bond Price-Yield Curve}
\usage{
qlg_bond_price_yield_curve(
  bond,
  ytm_center = NULL,
  spread = 0.02,
  n = 41,
  day_counter = Actual360(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\description{
Bond Price-Yield Curve
}
