% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_bond_sensitivity_table}
\alias{qlg_bond_sensitivity_table}
\title{Bond sensitivity table}
\usage{
qlg_bond_sensitivity_table(
  bond,
  ytm = NULL,
  shifts_bp = c(-100, -50, -25, -10, -1, 0, 1, 10, 25, 50, 100),
  day_counter = Actual365Fixed(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\arguments{
\item{bond}{A QuantLib bond object.}

\item{ytm}{Base yield. If NULL, calculated from bond.}

\item{shifts_bp}{Yield shifts in basis points.}

\item{day_counter}{QuantLib day counter.}

\item{compounding}{Compounding convention.}

\item{frequency}{Payment frequency.}
}
\value{
A tibble with price sensitivity to yield shifts.
}
\description{
Bond sensitivity table
}
