% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_bond_yield}
\alias{qlg_bond_yield}
\title{Bond Yield}
\usage{
qlg_bond_yield(
  bond,
  day_counter = Actual365Fixed(),
  compounding = "Compounded",
  frequency = "Annual"
)
}
\description{
Bond Yield
}
