% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_bond_yield_from_price}
\alias{qlg_bond_yield_from_price}
\title{Bond Yield from Clean Price}
\usage{
qlg_bond_yield_from_price(
  bond,
  clean_price,
  day_counter = Actual360(),
  compounding = "Compounded",
  frequency = "Annual",
  settlement_date = NULL
)
}
\description{
Bond Yield from Clean Price
}
