% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_bsm_process_bundle}
\alias{qlg_bsm_process_bundle}
\title{Build a Black-Scholes-Merton process bundle}
\usage{
qlg_bsm_process_bundle(
  valuation_date,
  spot,
  risk_free_rate,
  dividend_rate,
  volatility,
  day_counter = QuantLib::Actual365Fixed(),
  calendar = QuantLib::TARGET()
)
}
\arguments{
\item{valuation_date}{QuantLib Date.}

\item{spot}{Spot price.}

\item{risk_free_rate}{Risk-free rate.}

\item{dividend_rate}{Dividend yield.}

\item{volatility}{Volatility.}

\item{day_counter}{QuantLib day counter.}

\item{calendar}{QuantLib calendar.}
}
\value{
List with process and term structures.
}
\description{
Build a Black-Scholes-Merton process bundle
}
