% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_builder.R
\name{qlg_build_bond_discount_curve}
\alias{qlg_build_bond_discount_curve}
\title{Build a bond discount curve}
\usage{
qlg_build_bond_discount_curve(
  settlement_date = "2008-09-18",
  fixing_days = 3,
  settlement_days = 3
)
}
\arguments{
\item{settlement_date}{Settlement date.}

\item{fixing_days}{Fixing days for deposit helpers.}

\item{settlement_days}{Bond settlement days.}
}
\value{
A QuantLib yield term structure.
}
\description{
Build a bond discount curve
}
