% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_builder.R
\name{qlg_build_zero_curve}
\alias{qlg_build_zero_curve}
\title{Build a QuantLib zero curve from zero-rate nodes}
\usage{
qlg_build_zero_curve(date_chr, zero_rates, day_counter = "Actual365Fixed")
}
\arguments{
\item{date_chr}{Character vector of ISO dates.}

\item{zero_rates}{Numeric vector of zero rates.}

\item{day_counter}{Day counter name or QuantLib day counter object.}
}
\value{
A QuantLib ZeroCurve object.
}
\description{
Build a QuantLib zero curve from zero-rate nodes
}
