% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_callable_bond_dirty_price_safe}
\alias{qlg_callable_bond_dirty_price_safe}
\title{Safely obtain a callable bond dirty price}
\usage{
qlg_callable_bond_dirty_price_safe(bond)
}
\arguments{
\item{bond}{QuantLib callable bond.}
}
\value{
Numeric price or NA.
}
\description{
Safely obtain a callable bond dirty price
}
