% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_callable_bond_settlement_value_safe}
\alias{qlg_callable_bond_settlement_value_safe}
\title{Safely obtain a callable bond settlement value}
\usage{
qlg_callable_bond_settlement_value_safe(bond)
}
\arguments{
\item{bond}{QuantLib callable bond.}
}
\value{
Numeric value or NA.
}
\description{
Safely obtain a callable bond settlement value
}
