% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_cap_floor_optionlet_prices}
\alias{qlg_cap_floor_optionlet_prices}
\title{Extract cap/floor optionlet prices}
\usage{
qlg_cap_floor_optionlet_prices(cap_floor)
}
\description{
Extract cap/floor optionlet prices
}
