% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_cashflow_leg_summary}
\alias{qlg_cashflow_leg_summary}
\title{Summarise cashflows by leg}
\usage{
qlg_cashflow_leg_summary(cashflows)
}
\arguments{
\item{cashflows}{A cashflow schedule tibble.}
}
\value{
A tibble summarising cashflows by leg.
}
\description{
Summarise cashflows by leg
}
