% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_cashflow_schedule_from_trade}
\alias{qlg_cashflow_schedule_from_trade}
\title{Build cashflow schedule from one trade}
\usage{
qlg_cashflow_schedule_from_trade(trade, forecast_handle)
}
\arguments{
\item{trade}{A one-row trade data frame.}

\item{forecast_handle}{Forecast curve handle.}
}
\value{
A tibble containing cashflows.
}
\description{
Build cashflow schedule from one trade
}
