% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_cashflow_schedule_from_trades}
\alias{qlg_cashflow_schedule_from_trades}
\title{Build cashflow schedules from multiple trades}
\usage{
qlg_cashflow_schedule_from_trades(trades, forecast_handle)
}
\arguments{
\item{trades}{A trade data frame.}

\item{forecast_handle}{Forecast curve handle.}
}
\value{
A tibble containing cashflows for all supported trades.
}
\description{
Build cashflow schedules from multiple trades
}
