% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_cds_buyer_side}
\alias{qlg_cds_buyer_side}
\title{Obtain the QuantLib CDS buyer-side enum}
\usage{
qlg_cds_buyer_side()
}
\value{
QuantLib protection buyer enum.
}
\description{
Obtain the QuantLib CDS buyer-side enum
}
