% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_cds_cashflow_table}
\alias{qlg_cds_cashflow_table}
\title{Build a CDS cashflow and probability table}
\usage{
qlg_cds_cashflow_table(
  cds_schedule,
  protection_start_date,
  coupon_rate,
  hazard_curve,
  discount_curve,
  trade_date,
  notional
)
}
\arguments{
\item{cds_schedule}{QuantLib CDS schedule.}

\item{protection_start_date}{Protection start date.}

\item{coupon_rate}{Running coupon rate.}

\item{hazard_curve}{Hazard curve.}

\item{discount_curve}{Discount curve.}

\item{trade_date}{Trade date.}

\item{notional}{Notional.}
}
\value{
CDS cashflow table.
}
\description{
Build a CDS cashflow and probability table
}
