% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory.R
\name{qlg_cds_fair_spread}
\alias{qlg_cds_fair_spread}
\title{CDS fair spread}
\usage{
qlg_cds_fair_spread(cds)
}
\arguments{
\item{cds}{QuantLib CreditDefaultSwap object.}
}
\value{
Numeric fair spread.
}
\description{
CDS fair spread
}
