% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory.R
\name{qlg_cds_summary}
\alias{qlg_cds_summary}
\title{Summarise a CDS}
\usage{
qlg_cds_summary(cds)
}
\arguments{
\item{cds}{QuantLib CreditDefaultSwap object.}
}
\value{
A tibble with CDS analytics.
}
\description{
Summarise a CDS
}
