% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils.R
\name{qlg_chr}
\alias{qlg_chr}
\title{Convert QuantLib object to character}
\usage{
qlg_chr(x)
}
\arguments{
\item{x}{Object to print.}
}
\value{
Character scalar.
}
\description{
Convert QuantLib object to character
}
