% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_complete_accrual_metrics}
\alias{qlg_complete_accrual_metrics}
\title{Complete bond accrual metrics from a schedule}
\usage{
qlg_complete_accrual_metrics(
  price_tbl,
  bond_schedule,
  settlement_date,
  day_counter
)
}
\arguments{
\item{price_tbl}{Bond price metric table.}

\item{bond_schedule}{QuantLib Schedule.}

\item{settlement_date}{Settlement date.}

\item{day_counter}{QuantLib day counter.}
}
\value{
Updated metric table.
}
\description{
Complete bond accrual metrics from a schedule
}
