% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_curve_discount_safe}
\alias{qlg_curve_discount_safe}
\title{Safely obtain a discount factor}
\usage{
qlg_curve_discount_safe(curve, x)
}
\arguments{
\item{curve}{QuantLib yield curve.}

\item{x}{QuantLib Date or time.}
}
\value{
Numeric discount factor or NA.
}
\description{
Safely obtain a discount factor
}
