% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_helpers.R
\name{qlg_curve_table}
\alias{qlg_curve_table}
\title{Alias for qlg_curve_tbl}
\usage{
qlg_curve_table(
  curve,
  tenors = c("1D", "1W", "1M", "3M", "6M", "1Y", "2Y", "3Y", "5Y", "7Y", "10Y", "20Y",
    "30Y")
)
}
\description{
Alias for qlg_curve_tbl
}
