% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_helpers.R
\name{qlg_curve_tbl}
\alias{qlg_curve_tbl}
\title{Build a curve table}
\usage{
qlg_curve_tbl(
  curve,
  tenors = c("1D", "1W", "1M", "3M", "6M", "1Y", "2Y", "3Y", "5Y", "7Y", "10Y", "20Y",
    "30Y")
)
}
\arguments{
\item{curve}{QuantLib yield term structure.}

\item{tenors}{Character vector of tenors.}
}
\value{
Tibble with dates, discount factors, and zero rates.
}
\description{
Build a curve table
}
