% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{qlg_day_counter}
\alias{qlg_day_counter}
\title{Select QuantLib day counter}
\usage{
qlg_day_counter(day_counter = "Actual365Fixed")
}
\arguments{
\item{day_counter}{Day counter name or QuantLib day counter object.}
}
\value{
QuantLib day counter object.
}
\description{
Select QuantLib day counter
}
