% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{qlg_eval_date}
\alias{qlg_eval_date}
\title{Set QuantLib evaluation date}
\usage{
qlg_eval_date(x)
}
\arguments{
\item{x}{Date or character scalar.}
}
\value{
Invisibly returns QuantLib Date object.
}
\description{
Set QuantLib evaluation date
}
