% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_builder.R
\name{qlg_example_ois_nodes}
\alias{qlg_example_ois_nodes}
\title{Example OIS curve nodes}
\usage{
qlg_example_ois_nodes()
}
\value{
A tibble of sample zero-rate nodes.
}
\description{
Example OIS curve nodes
}
