% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_flat_curve}
\alias{qlg_flat_curve}
\title{Build a flat yield curve}
\usage{
qlg_flat_curve(
  reference_date,
  rate,
  day_counter = QuantLib::Actual365Fixed(),
  compounding = "Continuous"
)
}
\arguments{
\item{reference_date}{QuantLib Date.}

\item{rate}{Flat rate.}

\item{day_counter}{QuantLib day counter.}

\item{compounding}{Compounding convention.}
}
\value{
QuantLib FlatForward object.
}
\description{
Build a flat yield curve
}
