% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_helpers.R
\name{qlg_fmt_num}
\alias{qlg_fmt_num}
\title{Format a numeric value for display}
\usage{
qlg_fmt_num(x, digits = 6)
}
\arguments{
\item{x}{Numeric vector.}

\item{digits}{Number of digits.}
}
\value{
Character vector.
}
\description{
Format a numeric value for display
}
