% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_forward_swap_rate_from_schedule}
\alias{qlg_forward_swap_rate_from_schedule}
\title{Calculate a forward swap rate from a schedule}
\usage{
qlg_forward_swap_rate_from_schedule(schedule, curve, day_counter)
}
\arguments{
\item{schedule}{QuantLib Schedule.}

\item{curve}{QuantLib yield curve.}

\item{day_counter}{QuantLib day counter.}
}
\value{
Numeric forward swap rate.
}
\description{
Calculate a forward swap rate from a schedule
}
