% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_futures_bpv}
\alias{qlg_futures_bpv}
\title{Calculate futures BPV from CTD bond BPV}
\usage{
qlg_futures_bpv(ctd_bpv, conversion_factor, contracts = 1)
}
\arguments{
\item{ctd_bpv}{Numeric CTD bond BPV, DV01, or PV01.}

\item{conversion_factor}{Numeric futures conversion factor.}

\item{contracts}{Numeric number of futures contracts. Default is 1.}
}
\value{
Numeric futures BPV.
}
\description{
Calculate futures BPV using a practical CTD-based approximation:
}
\details{
\deqn{futures BPV = CTD bond BPV / conversion factor * contracts}
}
