% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_get_metric_num}
\alias{qlg_get_metric_num}
\title{Extract a numeric metric value}
\usage{
qlg_get_metric_num(tbl, metric_name)
}
\arguments{
\item{tbl}{Metric table.}

\item{metric_name}{Metric name.}
}
\value{
Numeric scalar or NA.
}
\description{
Extract a numeric metric value
}
