% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_hull_white_a}
\alias{qlg_hull_white_a}
\title{Calculate Hull-White A}
\usage{
qlg_hull_white_a(time, maturity, mean_reversion, sigma)
}
\arguments{
\item{time}{Start time.}

\item{maturity}{Maturity.}

\item{mean_reversion}{Mean reversion.}

\item{sigma}{Short-rate volatility.}
}
\value{
Numeric A.
}
\description{
Calculate Hull-White A
}
