% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_hull_white_b}
\alias{qlg_hull_white_b}
\title{Calculate Hull-White B}
\usage{
qlg_hull_white_b(time, maturity, mean_reversion)
}
\arguments{
\item{time}{Start time.}

\item{maturity}{Maturity.}

\item{mean_reversion}{Mean reversion.}
}
\value{
Numeric B.
}
\description{
Calculate Hull-White B
}
