% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_hull_white_convexity_bias}
\alias{qlg_hull_white_convexity_bias}
\title{Calculate Hull-White convexity bias}
\usage{
qlg_hull_white_convexity_bias(future_price, t, T, sigma = 0.01, a = 0.03)
}
\arguments{
\item{future_price}{Futures price.}

\item{t}{Time to futures maturity in years.}

\item{T}{Time to underlying maturity in years.}

\item{sigma}{Hull-White volatility.}

\item{a}{Hull-White mean reversion.}
}
\value{
Numeric convexity bias.
}
\description{
This is a thin wrapper around QuantLib HullWhite_convexityBias().
}
