% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_hull_white_model}
\alias{qlg_hull_white_model}
\title{Make a Hull-White one-factor short-rate model}
\usage{
qlg_hull_white_model(
  term_structure = NULL,
  valuation_date = qlg_eval_date_get(),
  rate = 0.03,
  a = 0.03,
  sigma = 0.01,
  day_counter = QuantLib::Actual365Fixed()
)
}
\arguments{
\item{term_structure}{Optional QuantLib yield term structure handle.}

\item{valuation_date}{Evaluation date used when term_structure is NULL.}

\item{rate}{Flat rate used when term_structure is NULL.}

\item{a}{Hull-White mean reversion.}

\item{sigma}{Hull-White volatility.}

\item{day_counter}{QuantLib day counter.}
}
\value{
QuantLib HullWhite model object.
}
\description{
Make a Hull-White one-factor short-rate model
}
