% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_hull_white_variance}
\alias{qlg_hull_white_variance}
\title{Calculate Hull-White variance term}
\usage{
qlg_hull_white_variance(time, maturity, mean_reversion, sigma)
}
\arguments{
\item{time}{Start time.}

\item{maturity}{Maturity.}

\item{mean_reversion}{Mean reversion.}

\item{sigma}{Short-rate volatility.}
}
\value{
Numeric variance term.
}
\description{
Calculate Hull-White variance term
}
