% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_hull_white_zero_rate}
\alias{qlg_hull_white_zero_rate}
\title{Calculate a Hull-White zero rate}
\usage{
qlg_hull_white_zero_rate(
  time,
  maturity,
  curve,
  mean_reversion,
  sigma,
  state_variable = 0
)
}
\arguments{
\item{time}{Start time.}

\item{maturity}{Maturity.}

\item{curve}{Initial yield curve.}

\item{mean_reversion}{Mean reversion.}

\item{sigma}{Short-rate volatility.}

\item{state_variable}{Hull-White state variable.}
}
\value{
Numeric zero rate.
}
\description{
Calculate a Hull-White zero rate
}
