% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_instrument_npv_safe}
\alias{qlg_instrument_npv_safe}
\title{Safely calculate an instrument NPV}
\usage{
qlg_instrument_npv_safe(instrument)
}
\arguments{
\item{instrument}{QuantLib instrument.}
}
\value{
Numeric NPV or NA.
}
\description{
Safely calculate an instrument NPV
}
