% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_quantive_riemann.R
\name{qlg_ir_basis_table}
\alias{qlg_ir_basis_table}
\title{Build a basis curve table}
\usage{
qlg_ir_basis_table(
  basis_env,
  tenors = c("1M", "3M", "6M", "1Y", "2Y", "3Y", "5Y", "10Y", "20Y", "30Y")
)
}
\description{
Build a basis curve table
}
