% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_quantive_riemann.R
\name{qlg_ir_make_basis_curve}
\alias{qlg_ir_make_basis_curve}
\title{Make a simple interpolated basis curve environment}
\usage{
qlg_ir_make_basis_curve(
  basis_data,
  base_curve_env,
  spread_label = "basis",
  verbose = TRUE
)
}
\description{
Make a simple interpolated basis curve environment
}
