% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_quantive_riemann.R
\name{qlg_leg_cashflow_at}
\alias{qlg_leg_cashflow_at}
\title{Safe accessor for a QuantLib cashflow leg}
\usage{
qlg_leg_cashflow_at(leg_obj, i_one_based)
}
\arguments{
\item{leg_obj}{QuantLib leg object.}

\item{i_one_based}{One-based cashflow index.}
}
\value{
A QuantLib cashflow object, or NULL if not accessible.
}
\description{
Safe accessor for a QuantLib cashflow leg
}
