% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_leg_cashflow_schedule}
\alias{qlg_leg_cashflow_schedule}
\title{Build cashflow schedule table from a QuantLib leg}
\usage{
qlg_leg_cashflow_schedule(leg, leg_name, leg_no = NA_integer_)
}
\arguments{
\item{leg}{A QuantLib Leg object.}

\item{leg_name}{Leg name.}

\item{leg_no}{Leg number. QuantLib uses 0-based leg indexing.}
}
\value{
A tibble containing cashflows for one leg.
}
\description{
Build cashflow schedule table from a QuantLib leg
}
