% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{qlg_leg_to_cashflow_tbl}
\alias{qlg_leg_to_cashflow_tbl}
\title{Convert QuantLib leg to cashflow table}
\usage{
qlg_leg_to_cashflow_tbl(leg)
}
\arguments{
\item{leg}{QuantLib Leg or CashFlow vector.}
}
\value{
A tibble with cashflow dates and amounts.
}
\description{
Convert QuantLib leg to cashflow table
}
