% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_lsm_american_put}
\alias{qlg_lsm_american_put}
\title{Approximate an American put with Longstaff-Schwartz}
\usage{
qlg_lsm_american_put(path_tbl, strike, risk_free_rate)
}
\arguments{
\item{path_tbl}{Long-form path table.}

\item{strike}{Strike.}

\item{risk_free_rate}{Continuously compounded discount rate.}
}
\value{
List with summary, exercise table, and diagnostics.
}
\description{
Approximate an American put with Longstaff-Schwartz
}
