% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics.R
\name{qlg_make_american_option}
\alias{qlg_make_american_option}
\title{Make a QuantLib American vanilla option}
\usage{
qlg_make_american_option(
  spot,
  strike,
  maturity_date,
  option_type = "put",
  valuation_date = qlg_eval_date_get(),
  risk_free_rate = 0.03,
  dividend_yield = 0,
  volatility = 0.2,
  day_counter = QuantLib::Actual365Fixed(),
  calendar = QuantLib::TARGET(),
  steps = 200L,
  pricing_engine = NULL
)
}
\arguments{
\item{spot}{Spot price.}

\item{strike}{Strike price.}

\item{maturity_date}{Option maturity date.}

\item{option_type}{Option type. Use "call" or "put".}

\item{valuation_date}{Evaluation date.}

\item{risk_free_rate}{Flat risk-free rate.}

\item{dividend_yield}{Flat dividend yield.}

\item{volatility}{Flat Black volatility.}

\item{day_counter}{QuantLib day counter.}

\item{calendar}{QuantLib calendar.}

\item{steps}{Number of binomial tree steps.}

\item{pricing_engine}{Optional QuantLib pricing engine.}
}
\value{
QuantLib VanillaOption object.
}
\description{
Make a QuantLib American vanilla option
}
