% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics.R
\name{qlg_make_american_option_from_trade}
\alias{qlg_make_american_option_from_trade}
\title{Make an American vanilla option from trade data}
\usage{
qlg_make_american_option_from_trade(trade, pricing_engine = NULL)
}
\arguments{
\item{trade}{A one-row data frame containing American option trade fields.}

\item{pricing_engine}{Optional QuantLib pricing engine.}
}
\value{
QuantLib VanillaOption object.
}
\description{
Make an American vanilla option from trade data
}
