% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics.R
\name{qlg_make_barrier_option}
\alias{qlg_make_barrier_option}
\title{Make a QuantLib barrier option}
\usage{
qlg_make_barrier_option(
  spot,
  strike,
  maturity_date,
  barrier,
  barrier_type = "down_out",
  rebate = 0,
  option_type = "call",
  valuation_date = qlg_eval_date_get(),
  risk_free_rate = 0.03,
  dividend_yield = 0,
  volatility = 0.2,
  day_counter = QuantLib::Actual365Fixed(),
  calendar = QuantLib::TARGET(),
  pricing_engine = NULL
)
}
\arguments{
\item{spot}{Spot price.}

\item{strike}{Strike price.}

\item{maturity_date}{Option maturity date.}

\item{barrier}{Barrier level.}

\item{barrier_type}{Barrier type. Use "down_out", "up_out", "down_in", or "up_in".}

\item{rebate}{Barrier rebate amount.}

\item{option_type}{Option type. Use "call" or "put".}

\item{valuation_date}{Evaluation date.}

\item{risk_free_rate}{Flat risk-free rate.}

\item{dividend_yield}{Flat dividend yield.}

\item{volatility}{Flat Black volatility.}

\item{day_counter}{QuantLib day counter.}

\item{calendar}{QuantLib calendar.}

\item{pricing_engine}{Optional QuantLib pricing engine.}
}
\value{
QuantLib BarrierOption object.
}
\description{
Make a QuantLib barrier option
}
