% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_make_bermudan_swaption}
\alias{qlg_make_bermudan_swaption}
\title{Make a QuantLib Bermudan swaption}
\usage{
qlg_make_bermudan_swaption(
  underlying_swap,
  exercise_dates,
  pricing_engine = NULL
)
}
\arguments{
\item{underlying_swap}{QuantLib VanillaSwap object.}

\item{exercise_dates}{Bermudan exercise dates.}

\item{pricing_engine}{Optional QuantLib pricing engine.}
}
\value{
QuantLib Swaption object.
}
\description{
Make a QuantLib Bermudan swaption
}
