% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_make_callability_schedule}
\alias{qlg_make_callability_schedule}
\title{Build a QuantLib callability schedule}
\usage{
qlg_make_callability_schedule(call_dates, call_price_clean = 100)
}
\arguments{
\item{call_dates}{Character, Date, or QuantLib dates.}

\item{call_price_clean}{Clean call price.}
}
\value{
QuantLib CallabilitySchedule.
}
\description{
Build a QuantLib callability schedule
}
