% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_make_cap_from_trade}
\alias{qlg_make_cap_from_trade}
\title{Make a cap from trade data}
\usage{
qlg_make_cap_from_trade(trade, pricing_engine = NULL)
}
\arguments{
\item{trade}{A one-row data frame containing cap trade fields.}

\item{pricing_engine}{Optional QuantLib pricing engine.}
}
\value{
QuantLib Cap object.
}
\description{
Make a cap from trade data
}
