% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{qlg_make_date_vector}
\alias{qlg_make_date_vector}
\title{Build QuantLib DateVector}
\usage{
qlg_make_date_vector(dates)
}
\arguments{
\item{dates}{Character vector or Date vector.}
}
\value{
QuantLib DateVector.
}
\description{
Build QuantLib DateVector
}
