% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory.R
\name{qlg_make_eonia_ois}
\alias{qlg_make_eonia_ois}
\title{Create an Eonia OIS swap}
\usage{
qlg_make_eonia_ois(swap_tenor, forecast_handle, fixed_rate)
}
\arguments{
\item{swap_tenor}{QuantLib Period object.}

\item{forecast_handle}{QuantLib yield term structure handle.}

\item{fixed_rate}{Fixed rate.}
}
\value{
A QuantLib OIS swap object.
}
\description{
Create an Eonia OIS swap
}
