% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_make_floor}
\alias{qlg_make_floor}
\title{Make a QuantLib interest-rate floor}
\usage{
qlg_make_floor(
  notional,
  start_date,
  maturity_date,
  floor_rate,
  valuation_date = qlg_eval_date_get(),
  tenor_months = 6L,
  discount_rate = 0.03,
  forecast_rate = discount_rate,
  volatility = 0.2,
  day_counter = QuantLib::Actual365Fixed(),
  calendar = QuantLib::TARGET(),
  convention = QuantLib::BusinessDayConvention_ModifiedFollowing_get(),
  pricing_engine = NULL
)
}
\description{
Make a QuantLib interest-rate floor
}
