% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_make_floor_from_trade}
\alias{qlg_make_floor_from_trade}
\title{Make a floor from trade data}
\usage{
qlg_make_floor_from_trade(trade, pricing_engine = NULL)
}
\arguments{
\item{trade}{A one-row data frame containing floor trade fields.}

\item{pricing_engine}{Optional QuantLib pricing engine.}
}
\value{
QuantLib Floor object.
}
\description{
Make a floor from trade data
}
