% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_make_interest_option_from_trade}
\alias{qlg_make_interest_option_from_trade}
\title{Make an interest-rate option from trade data}
\usage{
qlg_make_interest_option_from_trade(trade, pricing_engine = NULL)
}
\arguments{
\item{trade}{A one-row data frame containing interest-rate option trade fields.}

\item{pricing_engine}{Optional QuantLib pricing engine.}
}
\value{
QuantLib Cap or Floor object.
}
\description{
This dispatches to cap or floor creation depending on the trade fields.
}
