% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_make_normal_calculator}
\alias{qlg_make_normal_calculator}
\title{Build a normal-model pricing closure}
\usage{
qlg_make_normal_calculator(
  option_sign,
  strike,
  maturity,
  discount_factor,
  forward,
  vol
)
}
\arguments{
\item{option_sign}{1 for call, -1 for put.}

\item{strike}{Strike.}

\item{maturity}{Maturity.}

\item{discount_factor}{Discount factor.}

\item{forward}{Forward.}

\item{vol}{Normal volatility.}
}
\value{
Pricing function.
}
\description{
Build a normal-model pricing closure
}
