% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory.R
\name{qlg_make_ois}
\alias{qlg_make_ois}
\title{Create an OIS swap}
\usage{
qlg_make_ois(swap_tenor, overnight_index, fixed_rate)
}
\arguments{
\item{swap_tenor}{QuantLib Period object.}

\item{overnight_index}{QuantLib overnight index object.}

\item{fixed_rate}{Fixed rate.}
}
\value{
A QuantLib OIS swap object.
}
\description{
Create an OIS swap
}
