% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory.R
\name{qlg_make_ois_from_trade}
\alias{qlg_make_ois_from_trade}
\title{Make an OIS from trade data}
\usage{
qlg_make_ois_from_trade(trade, forecast_handle)
}
\arguments{
\item{trade}{A one-row data frame containing OIS trade fields.}

\item{forecast_handle}{QuantLib forecast curve handle.}
}
\value{
A QuantLib OIS object.
}
\description{
Make an OIS from trade data
}
