% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_make_path_generator}
\alias{qlg_make_path_generator}
\title{Build a QuantLib path generator}
\usage{
qlg_make_path_generator(
  process,
  maturity,
  n_steps,
  seed = 1L,
  sequence = c("pseudo", "sobol")
)
}
\arguments{
\item{process}{QuantLib stochastic process.}

\item{maturity}{Path maturity.}

\item{n_steps}{Number of time steps.}

\item{seed}{Random seed.}

\item{sequence}{Pseudo-random or Sobol.}
}
\value{
QuantLib path generator.
}
\description{
Build a QuantLib path generator
}
