% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_make_swaption}
\alias{qlg_make_swaption}
\title{Make a QuantLib swaption}
\usage{
qlg_make_swaption(underlying_swap, exercise_date, pricing_engine = NULL)
}
\arguments{
\item{underlying_swap}{QuantLib VanillaSwap object.}

\item{exercise_date}{Swaption exercise date.}

\item{pricing_engine}{Optional QuantLib pricing engine.}
}
\value{
QuantLib Swaption object.
}
\description{
Make a QuantLib swaption
}
