% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_make_swaption_from_trade}
\alias{qlg_make_swaption_from_trade}
\title{Make a Hull-White swaption from trade data}
\usage{
qlg_make_swaption_from_trade(
  trade,
  forecast_handle = NULL,
  discount_handle = NULL,
  pricing_engine = NULL
)
}
\arguments{
\item{trade}{A one-row data frame containing swaption trade fields.}

\item{forecast_handle}{Optional forecast curve handle.}

\item{discount_handle}{Optional discount curve handle.}

\item{pricing_engine}{Optional QuantLib swaption pricing engine.}
}
\value{
QuantLib Swaption object.
}
\description{
Make a Hull-White swaption from trade data
}
