% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory.R
\name{qlg_make_vanilla_swap}
\alias{qlg_make_vanilla_swap}
\title{Make a QuantLib VanillaSwap}
\usage{
qlg_make_vanilla_swap(
  effective_date,
  maturity_date,
  fixed_rate,
  notional = 1,
  forecast_handle = NULL,
  discount_handle = NULL,
  swap_type = "payer",
  index = "Euribor6M",
  spread = 0,
  fixed_tenor = QuantLib::Period(1, "Years"),
  floating_tenor = QuantLib::Period(6, "Months"),
  fixed_day_counter = QuantLib::Thirty360("European"),
  floating_day_counter = QuantLib::Actual360(),
  calendar = QuantLib::TARGET(),
  fixed_convention = "ModifiedFollowing",
  floating_convention = "ModifiedFollowing"
)
}
\arguments{
\item{effective_date}{Effective date.}

\item{maturity_date}{Maturity date.}

\item{fixed_rate}{Fixed coupon rate.}

\item{notional}{Notional amount.}

\item{forecast_handle}{Forecast curve handle.}

\item{discount_handle}{Discount curve handle. If supplied, a DiscountingSwapEngine is attached.}

\item{swap_type}{Payer or receiver.}

\item{index}{Floating index name.}

\item{spread}{Floating spread.}

\item{fixed_tenor}{Fixed leg tenor.}

\item{floating_tenor}{Floating leg tenor.}

\item{fixed_day_counter}{Fixed leg day counter.}

\item{floating_day_counter}{Floating leg day counter.}

\item{calendar}{QuantLib calendar.}

\item{fixed_convention}{Fixed leg business day convention.}

\item{floating_convention}{Floating leg business day convention.}
}
\value{
A QuantLib VanillaSwap object.
}
\description{
Make a QuantLib VanillaSwap
}
