% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_make_zero_spreaded_curve}
\alias{qlg_make_zero_spreaded_curve}
\title{Build a zero-spreaded yield curve}
\usage{
qlg_make_zero_spreaded_curve(
  base_curve_handle,
  spread_rate,
  day_counter = QuantLib::Actual365Fixed(),
  compounding = QuantLib::Compounding_Simple_get(),
  frequency = QuantLib::Frequency_Annual_get(),
  extrapolate = TRUE
)
}
\arguments{
\item{base_curve_handle}{QuantLib yield curve handle.}

\item{spread_rate}{Zero-rate spread.}

\item{day_counter}{QuantLib day counter.}

\item{compounding}{QuantLib compounding convention.}

\item{frequency}{QuantLib frequency convention.}

\item{extrapolate}{Enable extrapolation.}
}
\value{
List containing curve, curve_handle, and spread_handle.
}
\description{
Build a zero-spreaded yield curve
}
