% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_metric_tbl_display}
\alias{qlg_metric_tbl_display}
\title{Format a metric table for display}
\usage{
qlg_metric_tbl_display(tbl)
}
\arguments{
\item{tbl}{Metric table.}
}
\value{
A tibble with a character value column when present.
}
\description{
Format a metric table for display
}
