% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_metric_value_to_chr}
\alias{qlg_metric_value_to_chr}
\title{Convert a metric value to character}
\usage{
qlg_metric_value_to_chr(x)
}
\arguments{
\item{x}{Value or list-column element.}
}
\value{
Character scalar.
}
\description{
Convert a metric value to character
}
