% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_mid_date}
\alias{qlg_mid_date}
\title{Calculate a midpoint date}
\usage{
qlg_mid_date(first_date, second_date)
}
\arguments{
\item{first_date}{First QuantLib date.}

\item{second_date}{Second QuantLib date.}
}
\value{
QuantLib midpoint date.
}
\description{
Calculate a midpoint date
}
