% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_normal_vol_from_price}
\alias{qlg_normal_vol_from_price}
\title{Infer normal volatility from an option price}
\usage{
qlg_normal_vol_from_price(
  option_sign,
  strike,
  forward,
  maturity,
  option_npv,
  discount_factor
)
}
\arguments{
\item{option_sign}{1 for call, -1 for put.}

\item{strike}{Strike.}

\item{forward}{Forward.}

\item{maturity}{Maturity.}

\item{option_npv}{Option NPV.}

\item{discount_factor}{Discount factor.}
}
\value{
Normal volatility.
}
\description{
Infer normal volatility from an option price
}
