% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_normal_vol_hagan}
\alias{qlg_normal_vol_hagan}
\title{Calculate Hagan normal SABR volatility}
\usage{
qlg_normal_vol_hagan(strike, forward, maturity, beta, alpha, volvol, rho)
}
\arguments{
\item{strike}{Strike.}

\item{forward}{Forward.}

\item{maturity}{Maturity.}

\item{beta}{SABR beta.}

\item{alpha}{SABR alpha.}

\item{volvol}{SABR volatility of volatility.}

\item{rho}{SABR correlation.}
}
\value{
Numeric normal volatility.
}
\description{
Calculate Hagan normal SABR volatility
}
