% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_ois_cashflow_schedule}
\alias{qlg_ois_cashflow_schedule}
\title{Build OIS cashflow schedule table from fixed and overnight legs}
\usage{
qlg_ois_cashflow_schedule(swap)
}
\arguments{
\item{swap}{A QuantLib OIS-like object.}
}
\value{
A tibble containing fixed and overnight leg cashflows.
}
\description{
Build OIS cashflow schedule table from fixed and overnight legs
}
