% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_ois_cashflow_schedule_from_trade}
\alias{qlg_ois_cashflow_schedule_from_trade}
\title{Build an OIS cashflow schedule from trade data}
\usage{
qlg_ois_cashflow_schedule_from_trade(trade, forecast_handle)
}
\arguments{
\item{trade}{A one-row data frame containing OIS trade fields.}

\item{forecast_handle}{QuantLib forecast curve handle.}
}
\value{
A tibble containing fixed and overnight leg cashflows.
}
\description{
Build an OIS cashflow schedule from trade data
}
