% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_ois_overnight_leg_table}
\alias{qlg_ois_overnight_leg_table}
\title{OIS overnight leg cashflow table}
\usage{
qlg_ois_overnight_leg_table(swap)
}
\arguments{
\item{swap}{A QuantLib OIS-like object.}
}
\value{
A tibble of overnight leg cashflows.
}
\description{
OIS overnight leg cashflow table
}
