% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_ois_summary}
\alias{qlg_ois_summary}
\title{Summarise OIS cashflow legs}
\usage{
qlg_ois_summary(fixed_leg, overnight_leg)
}
\arguments{
\item{fixed_leg}{Fixed leg cashflow table.}

\item{overnight_leg}{Overnight leg cashflow table.}
}
\value{
A tibble summarising OIS cashflow legs.
}
\description{
Summarise OIS cashflow legs
}
