% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics.R
\name{qlg_option_delta}
\alias{qlg_option_delta}
\title{Option delta}
\usage{
qlg_option_delta(option)
}
\arguments{
\item{option}{QuantLib option object.}
}
\value{
Numeric delta.
}
\description{
Option delta
}
