% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics.R
\name{qlg_option_gamma}
\alias{qlg_option_gamma}
\title{Option gamma}
\usage{
qlg_option_gamma(option)
}
\arguments{
\item{option}{QuantLib option object.}
}
\value{
Numeric gamma.
}
\description{
Option gamma
}
