% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_option_metric_table}
\alias{qlg_option_metric_table}
\title{Build an option metric table}
\usage{
qlg_option_metric_table(option, process = NULL)
}
\arguments{
\item{option}{QuantLib option.}

\item{process}{Optional pricing process.}
}
\value{
Metric tibble.
}
\description{
Build an option metric table
}
