% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics.R
\name{qlg_option_summary}
\alias{qlg_option_summary}
\title{Summarise an option}
\usage{
qlg_option_summary(option)
}
\arguments{
\item{option}{QuantLib option object.}
}
\value{
A tibble with option analytics.
}
\description{
Summarise an option
}
