% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics.R
\name{qlg_option_theta}
\alias{qlg_option_theta}
\title{Option theta}
\usage{
qlg_option_theta(option)
}
\arguments{
\item{option}{QuantLib option object.}
}
\value{
Numeric theta.
}
\description{
Option theta
}
