% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_analytics.R
\name{qlg_option_vega}
\alias{qlg_option_vega}
\title{Option vega}
\usage{
qlg_option_vega(option)
}
\arguments{
\item{option}{QuantLib option object.}
}
\value{
Numeric vega.
}
\description{
Option vega
}
