% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_helpers.R
\name{qlg_path_tbl}
\alias{qlg_path_tbl}
\title{Convert path-like simulation output to a tibble}
\usage{
qlg_path_tbl(path)
}
\arguments{
\item{path}{Path object, matrix, data frame, or numeric vector.}
}
\value{
Tibble.
}
\description{
Convert path-like simulation output to a tibble
}
