% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils.R
\name{qlg_period_chr}
\alias{qlg_period_chr}
\title{Build and print a QuantLib Period}
\usage{
qlg_period_chr(x)
}
\arguments{
\item{x}{Tenor string such as 1D, 1W, 3M, or 2Y.}
}
\value{
Character scalar.
}
\description{
Build and print a QuantLib Period
}
