% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils.R
\name{qlg_period_days}
\alias{qlg_period_days}
\title{Build a QuantLib day Period}
\usage{
qlg_period_days(n)
}
\arguments{
\item{n}{Number of days.}
}
\value{
A QuantLib Period object.
}
\description{
Build a QuantLib day Period
}
