% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils.R
\name{qlg_period_months}
\alias{qlg_period_months}
\title{Build a QuantLib month Period}
\usage{
qlg_period_months(n)
}
\arguments{
\item{n}{Number of months.}
}
\value{
A QuantLib Period object.
}
\description{
Build a QuantLib month Period
}
