% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils.R
\name{qlg_period_weeks}
\alias{qlg_period_weeks}
\title{Build a QuantLib week Period}
\usage{
qlg_period_weeks(n)
}
\arguments{
\item{n}{Number of weeks.}
}
\value{
A QuantLib Period object.
}
\description{
Build a QuantLib week Period
}
