% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils.R
\name{qlg_period_years}
\alias{qlg_period_years}
\title{Build a QuantLib year Period}
\usage{
qlg_period_years(n)
}
\arguments{
\item{n}{Number of years.}
}
\value{
A QuantLib Period object.
}
\description{
Build a QuantLib year Period
}
