% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_push_calibration_helpers}
\alias{qlg_push_calibration_helpers}
\title{Push calibration helpers into a QuantLib vector}
\usage{
qlg_push_calibration_helpers(helpers)
}
\arguments{
\item{helpers}{List of QuantLib calibration helpers.}
}
\value{
QuantLib CalibrationHelperVector.
}
\description{
Push calibration helpers into a QuantLib vector
}
