% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{qlg_push_rate_helpers}
\alias{qlg_push_rate_helpers}
\title{Push rate helpers into QuantLib RateHelperVector}
\usage{
qlg_push_rate_helpers(helpers)
}
\arguments{
\item{helpers}{List of QuantLib rate helpers.}
}
\value{
QuantLib RateHelperVector.
}
\description{
Push rate helpers into QuantLib RateHelperVector
}
