% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{qlg_quote_handle}
\alias{qlg_quote_handle}
\title{Create QuantLib QuoteHandle}
\usage{
qlg_quote_handle(x)
}
\arguments{
\item{x}{Numeric quote.}
}
\value{
QuantLib QuoteHandle.
}
\description{
Create QuantLib QuoteHandle
}
