% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_replace_metric_value}
\alias{qlg_replace_metric_value}
\title{Replace or append a metric value}
\usage{
qlg_replace_metric_value(tbl, metric_name, new_value)
}
\arguments{
\item{tbl}{Metric table.}

\item{metric_name}{Metric name.}

\item{new_value}{Replacement value.}
}
\value{
Updated metric table.
}
\description{
Replace or append a metric value
}
