% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_rmse}
\alias{qlg_rmse}
\title{Calculate root mean squared error}
\usage{
qlg_rmse(actual, fitted)
}
\arguments{
\item{actual}{Actual values.}

\item{fitted}{Fitted values.}
}
\value{
Numeric RMSE.
}
\description{
Calculate root mean squared error
}
