% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_safe_iso}
\alias{qlg_safe_iso}
\title{Safely convert a QuantLib date to ISO}
\usage{
qlg_safe_iso(x)
}
\arguments{
\item{x}{QuantLib Date object or NULL.}
}
\value{
ISO date string or NA.
}
\description{
Safely convert a QuantLib date to ISO
}
