% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_safe_num}
\alias{qlg_safe_num}
\title{Safely convert a value to numeric}
\usage{
qlg_safe_num(x)
}
\arguments{
\item{x}{Object convertible to numeric.}
}
\value{
Numeric scalar or NA.
}
\description{
Safely convert a value to numeric
}
