% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_safe_sabr_vol}
\alias{qlg_safe_sabr_vol}
\title{Safely calculate Black SABR volatility}
\usage{
qlg_safe_sabr_vol(strike, forward, maturity, alpha, beta, volvol, rho)
}
\arguments{
\item{strike}{Strike.}

\item{forward}{Forward.}

\item{maturity}{Maturity.}

\item{alpha}{SABR alpha.}

\item{beta}{SABR beta.}

\item{volvol}{SABR volatility of volatility.}

\item{rho}{SABR correlation.}
}
\value{
Numeric volatility or NA.
}
\description{
Safely calculate Black SABR volatility
}
