% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_schedule_date_vector}
\alias{qlg_schedule_date_vector}
\title{Build a list of dates from a QuantLib schedule}
\usage{
qlg_schedule_date_vector(schedule)
}
\arguments{
\item{schedule}{QuantLib Schedule object.}
}
\value{
List of QuantLib Date objects.
}
\description{
Build a list of dates from a QuantLib schedule
}
