% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils.R
\name{qlg_schedule_dates}
\alias{qlg_schedule_dates}
\title{Build a table of QuantLib schedule dates}
\usage{
qlg_schedule_dates(schedule)
}
\arguments{
\item{schedule}{QuantLib Schedule object.}
}
\value{
A tibble with schedule_date.
}
\description{
Build a table of QuantLib schedule dates
}
