% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_schedule_table}
\alias{qlg_schedule_table}
\title{Build a schedule table with period numbers}
\usage{
qlg_schedule_table(schedule)
}
\arguments{
\item{schedule}{QuantLib Schedule object.}
}
\value{
Tibble with period and schedule_date.
}
\description{
Build a schedule table with period numbers
}
