% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_set_bond_pricing_engine}
\alias{qlg_set_bond_pricing_engine}
\title{Create a zero coupon bond}
\usage{
qlg_set_bond_pricing_engine(bond, curve)
}
\arguments{
\item{bond}{QuantLib bond object.}

\item{curve}{QuantLib yield term structure.}
}
\value{
Bond with pricing engine.
}
\description{
Create a zero coupon bond
}
