% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_period_utils.R
\name{qlg_set_eval_date}
\alias{qlg_set_eval_date}
\title{Set QuantLib evaluation date}
\usage{
qlg_set_eval_date(eval_date)
}
\arguments{
\item{eval_date}{ISO date string, R Date, or QuantLib Date.}
}
\value{
The QuantLib Date invisibly.
}
\description{
Set QuantLib evaluation date
}
