% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_shifted_normal_vol_hagan}
\alias{qlg_shifted_normal_vol_hagan}
\title{Calculate shifted Hagan normal SABR volatility}
\usage{
qlg_shifted_normal_vol_hagan(
  strike,
  forward,
  maturity,
  beta,
  alpha,
  volvol,
  rho,
  shift = 0.025
)
}
\arguments{
\item{strike}{Strike.}

\item{forward}{Forward.}

\item{maturity}{Maturity.}

\item{beta}{SABR beta.}

\item{alpha}{SABR alpha.}

\item{volvol}{SABR volatility of volatility.}

\item{rho}{SABR correlation.}

\item{shift}{Shift applied to strike and forward.}
}
\value{
Numeric normal volatility or NA.
}
\description{
Calculate shifted Hagan normal SABR volatility
}
