% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_swap_annuity_from_schedule}
\alias{qlg_swap_annuity_from_schedule}
\title{Calculate swap annuity from a schedule}
\usage{
qlg_swap_annuity_from_schedule(schedule, curve, day_counter)
}
\arguments{
\item{schedule}{QuantLib Schedule.}

\item{curve}{QuantLib yield curve.}

\item{day_counter}{QuantLib day counter.}
}
\value{
Numeric annuity.
}
\description{
Calculate swap annuity from a schedule
}
