% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_swap_fair_rate}
\alias{qlg_swap_fair_rate}
\title{Swap fair fixed rate}
\usage{
qlg_swap_fair_rate(swap)
}
\arguments{
\item{swap}{A QuantLib VanillaSwap-like object.}
}
\description{
Swap fair fixed rate
}
