% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_swap_fair_spread}
\alias{qlg_swap_fair_spread}
\title{Swap fair spread}
\usage{
qlg_swap_fair_spread(swap)
}
\arguments{
\item{swap}{A QuantLib VanillaSwap-like object.}
}
\description{
Swap fair spread
}
