% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_swap_fixed_leg_npv}
\alias{qlg_swap_fixed_leg_npv}
\title{Swap fixed leg NPV}
\usage{
qlg_swap_fixed_leg_npv(swap)
}
\arguments{
\item{swap}{A QuantLib VanillaSwap-like object.}
}
\description{
Swap fixed leg NPV
}
