% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_swap_fixed_leg_table}
\alias{qlg_swap_fixed_leg_table}
\title{Swap fixed leg cashflow table}
\usage{
qlg_swap_fixed_leg_table(swap)
}
\arguments{
\item{swap}{A QuantLib VanillaSwap-like object.}
}
\value{
A tibble of fixed leg cashflows.
}
\description{
Swap fixed leg cashflow table
}
