% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_swap_leg_npv}
\alias{qlg_swap_leg_npv}
\title{Swap leg NPV}
\usage{
qlg_swap_leg_npv(swap, leg_no)
}
\arguments{
\item{swap}{A QuantLib Swap object.}

\item{leg_no}{Leg number. QuantLib uses 0-based leg indexing.}
}
\value{
Swap leg NPV.
}
\description{
Swap leg NPV
}
