% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_swaption_annuity}
\alias{qlg_swaption_annuity}
\title{Calculate swaption annuity}
\usage{
qlg_swaption_annuity(swaption)
}
\arguments{
\item{swaption}{QuantLib Swaption object.}
}
\value{
Numeric annuity if available.
}
\description{
Calculate swaption annuity
}
