% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_swaption_npv}
\alias{qlg_swaption_npv}
\title{Calculate swaption NPV}
\usage{
qlg_swaption_npv(swaption)
}
\arguments{
\item{swaption}{QuantLib Swaption object.}
}
\value{
Numeric NPV.
}
\description{
Calculate swaption NPV
}
