% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_swaption_summary}
\alias{qlg_swaption_summary}
\title{Summarise a swaption}
\usage{
qlg_swaption_summary(swaption)
}
\arguments{
\item{swaption}{QuantLib Swaption object.}
}
\value{
A tibble with swaption analytics.
}
\description{
Summarise a swaption
}
