% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_option.R
\name{qlg_swaption_vega}
\alias{qlg_swaption_vega}
\title{Calculate swaption vega}
\usage{
qlg_swaption_vega(swaption)
}
\arguments{
\item{swaption}{QuantLib Swaption object.}
}
\value{
Numeric vega if available.
}
\description{
Calculate swaption vega
}
