% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_quantive_riemann.R
\name{qlg_trade_ois_swap}
\alias{qlg_trade_ois_swap}
\title{Build an OIS swap trade from a curve environment}
\usage{
qlg_trade_ois_swap(
  curve_env,
  effective = NULL,
  maturity,
  nominal = 1e+06,
  fixed_rate,
  float_spread = 0,
  swap_type = QuantLib::Swap_Payer_get(),
  fixed_bdc = "ModifiedFollowing",
  date_rule = "Backward",
  eom = FALSE,
  fixed_schedule_tenor = NULL,
  overnight_schedule_tenor = NULL,
  payment_lag = NULL,
  verbose = TRUE
)
}
\arguments{
\item{curve_env}{Curve environment list.}

\item{effective}{Optional effective date.}

\item{maturity}{Maturity date.}

\item{nominal}{Nominal amount.}

\item{fixed_rate}{Fixed rate.}

\item{float_spread}{Floating spread.}

\item{swap_type}{QuantLib swap type.}

\item{fixed_bdc}{Business day convention.}

\item{date_rule}{Date generation rule.}

\item{eom}{End-of-month flag.}

\item{fixed_schedule_tenor}{Optional fixed leg schedule tenor.}

\item{overnight_schedule_tenor}{Optional overnight leg schedule tenor.}

\item{payment_lag}{Optional payment lag.}

\item{verbose}{Whether to print progress messages.}
}
\value{
Trade environment list.
}
\description{
Build an OIS swap trade from a curve environment
}
