% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quantlib.R
\name{qlg_use_quantlib}
\alias{qlg_use_quantlib}
\title{Use local QuantLib installation}
\usage{
qlg_use_quantlib(
  path = "C:/AnalyticFin/Projects/QuantLib_tidy/library/windows"
)
}
\description{
Use local QuantLib installation
}
