% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_analytics.R
\name{qlg_value_cashflow_schedule_from_trades}
\alias{qlg_value_cashflow_schedule_from_trades}
\title{Value cashflow schedules from multiple trades}
\usage{
qlg_value_cashflow_schedule_from_trades(
  trades,
  forecast_handle,
  discount_curve,
  fixings = NULL,
  evaluation_date = NULL
)
}
\arguments{
\item{trades}{A trade data frame.}

\item{forecast_handle}{Forecast curve handle.}

\item{discount_curve}{Discount curve or discount curve handle.}

\item{fixings}{Optional fixing tibble.}

\item{evaluation_date}{Evaluation date.}
}
\value{
A tibble with cashflows, discount factors, fixings, forwards, and PVs.
}
\description{
Value cashflow schedules from multiple trades
}
