% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_vasicek_a}
\alias{qlg_vasicek_a}
\title{Calculate Vasicek A}
\usage{
qlg_vasicek_a(time, maturity, mean_reversion, sigma, long_run_rate)
}
\arguments{
\item{time}{Start time.}

\item{maturity}{Maturity.}

\item{mean_reversion}{Mean reversion.}

\item{sigma}{Short-rate volatility.}

\item{long_run_rate}{Long-run rate.}
}
\value{
Numeric A.
}
\description{
Calculate Vasicek A
}
