% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_vasicek_b}
\alias{qlg_vasicek_b}
\title{Calculate Vasicek B}
\usage{
qlg_vasicek_b(time, maturity, mean_reversion)
}
\arguments{
\item{time}{Start time.}

\item{maturity}{Maturity.}

\item{mean_reversion}{Mean reversion.}
}
\value{
Numeric B.
}
\description{
Calculate Vasicek B
}
