% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_BQHs.R
\name{qlg_vasicek_zero_rate}
\alias{qlg_vasicek_zero_rate}
\title{Calculate a Vasicek zero rate}
\usage{
qlg_vasicek_zero_rate(
  time,
  maturity,
  mean_reversion,
  sigma,
  long_run_rate,
  short_rate
)
}
\arguments{
\item{time}{Start time.}

\item{maturity}{Maturity.}

\item{mean_reversion}{Mean reversion.}

\item{sigma}{Short-rate volatility.}

\item{long_run_rate}{Long-run rate.}

\item{short_rate}{Current short rate.}
}
\value{
Numeric zero rate.
}
\description{
Calculate a Vasicek zero rate
}
