% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_builder.R
\name{qlg_yield_curve_handle}
\alias{qlg_yield_curve_handle}
\title{Build a QuantLib yield curve handle}
\usage{
qlg_yield_curve_handle(curve)
}
\arguments{
\item{curve}{A QuantLib yield term structure object.}
}
\value{
A QuantLib YieldTermStructureHandle object.
}
\description{
Build a QuantLib yield curve handle
}
