% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_zero_coupon_bond}
\alias{qlg_zero_coupon_bond}
\title{Get bond coupon information}
\usage{
qlg_zero_coupon_bond(
  discount_curve,
  maturity = "2013-08-15",
  issue = "2008-08-15",
  face = 100
)
}
\arguments{
\item{bond}{A QuantLib bond object.}

\item{as_of}{Evaluation date. Default is qlg_get_eval_date().}
}
\value{
A tibble with coupon cashflow information.
}
\description{
Get bond coupon information
}
