% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_bond_settlement_info}
\alias{qlg_bond_settlement_info}
\title{Get bond settlement information}
\usage{
qlg_bond_settlement_info(bond, as_of = qlg_get_eval_date())
}
\description{
Get bond settlement information
}
