% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflows.R
\name{qlg_build_bond_discount_curve}
\alias{qlg_build_bond_discount_curve}
\title{Build a bond discount curve}
\usage{
qlg_build_bond_discount_curve(
  settlement_date = "2008-09-18",
  fixing_days = 3,
  settlement_days = 3
)
}
\description{
Build a bond discount curve
}
