% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflows.R
\name{qlg_build_discount_curve}
\alias{qlg_build_discount_curve}
\title{Build a QuantLib discount curve from zero-rate nodes}
\usage{
qlg_build_discount_curve(nodes, day_counter = "Actual365Fixed")
}
\description{
Build a QuantLib discount curve from zero-rate nodes
}
