% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_fixed_rate_bond}
\alias{qlg_fixed_rate_bond}
\title{Create Fixed Rate Bond}
\usage{
qlg_fixed_rate_bond(
  issue_date = "2007-05-15",
  maturity_date = "2017-05-15",
  coupon_rate = 0.045,
  face_amount = 100,
  settlement_days = 3
)
}
\description{
Create Fixed Rate Bond
}
