% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_factory.R
\name{qlg_make_asset_swap}
\alias{qlg_make_asset_swap}
\title{Build a QuantLib AssetSwap}
\usage{
qlg_make_asset_swap(
  bond,
  clean_price,
  ibor_index,
  floating_schedule,
  floating_day_counter,
  deal_maturity,
  pay_bond_coupon = TRUE,
  spread = 0,
  par_asset_swap = TRUE,
  gearing = 1,
  non_par_repayment = 100
)
}
\arguments{
\item{bond}{QuantLib bond object.}

\item{clean_price}{Bond clean price.}

\item{ibor_index}{QuantLib IborIndex object.}

\item{floating_schedule}{QuantLib Schedule object.}

\item{floating_day_counter}{QuantLib DayCounter object.}

\item{deal_maturity}{QuantLib Date object.}

\item{pay_bond_coupon}{Logical. TRUE means paying the bond coupon leg.}

\item{spread}{Floating leg spread.}

\item{par_asset_swap}{Logical. TRUE for par asset swap.}

\item{gearing}{Floating leg gearing.}

\item{non_par_repayment}{Non-par repayment amount.}
}
\value{
QuantLib AssetSwap object.
}
\description{
This is a thin wrapper around QuantLib AssetSwap.
}
