% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_analytics.R
\name{qlg_zero_coupon_bond}
\alias{qlg_zero_coupon_bond}
\title{Create a zero coupon bond}
\usage{
qlg_zero_coupon_bond(
  discount_curve,
  maturity = "2013-08-15",
  issue = "2008-08-15",
  face = 100
)
}
\description{
Create a zero coupon bond
}
